fix(RatingCounter): 계산 과정에서 int overflow 발생 부분 수정

+++ src/main/java/lotto/domain/rating/RatingCounter.java
@@ -27,8 +27,8 @@ public class RatingCounter {
     public double totalSum() {
         double sum = 0;
         for (Rating rating : Rating.values()) {
-            sum += rating.getReward() * ratings.get(rating)
-                                               .getCount();
+            sum += (double) rating.getReward() * ratings.get(rating)
+                                                        .getCount();
         }
         return sum;
     }
